requirements.txt
streamlit==1.52.2
google-generativeai==0.8.3
python-dotenv==1.0.0
requests==2.31.0
st-annotated-text==4.0.1
PyPDF2
python-docx
Pillow
pytesseract
pandas
numpy
scikit-learn
matplotlib
seaborn
nltk